Draper changes (#13)
* Linking against VmfTestLib

removed the install for yaml-cpp in the dockerfile and pointed cmake at vmf's yaml-cpp instead

* Removed irrelevant modules

* Changed rand argument types from size_t to ul

Arguments to vmf::VmfRand::randBetween are now casted to (or declared as) unsigned long before being passed. The result is casted back to size_t to minimize codebase changes.

* Updated "Last Updated" column

+++ Radamsa/vmf/src/modules/common/mutator/RadamsaDecrementByteMutator.cpp
@@ -139,15 +139,15 @@ void RadamsaDecrementByteMutator::mutateTestCase(StorageModule& storage, Storage
     memcpy(newBuffer, originalBuffer, originalSize);
 
     // Select a random byte to circularly decrement.
-    const size_t lower{0u};
+    const unsigned long lower{0ul};
     const size_t upper{originalSize - 1u};
-    const size_t maximumRandomIndexValue{originalSize - minimumSeedIndex};
+    const unsigned long maximumRandomIndexValue{static_cast<unsigned long>(originalSize - minimumSeedIndex)};
     const size_t randomIndexToDecrement{
                                     std::clamp(
-                                        rand->randBetween(
+                                        static_cast<size_t>(rand->randBetween(
                                             lower,
-                                            maximumRandomIndexValue) + minimumSeedIndex,
-                                        lower,
+                                            maximumRandomIndexValue)) + minimumSeedIndex,
+                                        static_cast<size_t>(lower),
                                         upper
                                     )
     };
