+++ Radamsa/vmf/src/modules/common/mutator/RadamsaDropByteMutator.cpp
@@ -141,15 +141,15 @@ void RadamsaDropByteMutator::mutateTestCase(StorageModule& storage, StorageEntry
 
     // Select a random byte to drop from the original buffer.
 
-    const size_t lower{0u};
+    const unsigned long lower{0ul};
     const size_t upper{originalSize - 1u};
-    const size_t maximumRandomIndexValue{originalSize - minimumSeedIndex};
+    const unsigned long maximumRandomIndexValue{static_cast<unsigned long>(originalSize - minimumSeedIndex)};
     const size_t randomIndexToDrop{
                             std::clamp(
-                                    rand->randBetween(
+                                    static_cast<size_t>(rand->randBetween(
                                                         lower,
-                                                        maximumRandomIndexValue) + minimumSeedIndex,
-                                                        lower,
+                                                        maximumRandomIndexValue)) + minimumSeedIndex,
+                                                        static_cast<size_t>(lower),
                                                         upper)};
 
     // Copy data from the original buffer into the new buffer, but exclude the random byte.
